requirements.txt
Pillow==5.1.0
Flask==1.0.2
requests==2.18.4
numpy==1.14.3